ai-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
langchain==0.3.0
langchain-openai==0.2.0
langchain-core==0.3.0
supabase==2.7.4
python-jose[cryptography]==3.3.0
resend==2.0.0
apscheduler==3.10.4
pydantic==2.9.2
httpx==0.27.2
langchain-google-genai==4.2.1
groq==1.1.2
playwright==1.47.0
